+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+# [0.2.0] - 10/07/2023
+
+- Added a `Parser#find_by_tag_names(['tag-name'])` method to the parser that returns a Hash keyed off the tag name.
+- Removed `tag_names` as a parameters for `Parser#find_custom_elements`
+
 ## [0.1.0] - 2023-10-03
 
 - Initial release